Refactor agent code for readability
